cuoti-system/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
qrcode[pil]==7.4.2
reportlab==4.0.0
python-docx==1.1.0
Pillow==10.1.0
pdfplumber==0.10.0
openpyxl==3.1.0
